+++ inst/testme/test-futurize-mgcv.R
@@ -17,5 +17,20 @@ print(b)
 
 stopifnot(all.equal(summary(b), summary(b_truth)))
 
+## predict.bam() via stats::predict() S3 generic dispatch
+stopifnot(inherits(b_truth, "bam"))
+
+newdat <- dat[1:100, ]
+p_truth <- predict(b_truth, newdata = newdat)
+
+## mgcv exports predict.bam(), so we can call it explictly
+p0 <- predict.bam(b_truth, newdata = newdat) |> futurize()
+stopifnot(all.equal(p0, p_truth))
+
+## futurize() when dispatching to S3 method predict() for 'bam'
+p <- predict(b_truth, newdata = newdat) |> futurize()
+stopifnot(all.equal(p, p_truth))
+
+
 plan(sequential)
 } ## if (requireNamespace("mgcv"))